requirements.txt
fastapi
uvicorn
redis
bcrypt
pillow
tensorflow
numpy
pydantic
scikit-learn
python-multipart
streamlit
requests
opencv-python